asm-generic: don't #include <linux/atomic.h> in qspinlock_types.h
The "qspinlock_types.h" doesn't need linux/atomic.h directly. And
because of this, and because including it requires the protection
against recursive inclusion, it looks reasonable to move the
inclusion exactly where it is needed. This change affects the x86
arch, as the only user of qspinlocks at now. I have build-tested
the change on x86 with CONFIG_PARAVIRT enabled and disabled.

Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>