module: verify address is read-only
Patch series "provide check for ro_after_init memory sections", v5.

Provide a mechanism for other functions to verify that their arguments are
read-only.

This implements the first half of a suggestion made by Kees Cook for the
Kernel Self Protection Project:

    - provide mechanism to check for ro_after_init memory areas, and
      reject structures not marked ro_after_init in vmbus_register()

      http://www.openwall.com/lists/kernel-hardening/2017/02/04/1

The idea is to prevent structures (including modules) that are not
read-only from being passed to functions.  It builds upon the functions in
kernel/extable.c that test if an address is in the text section.

A build failure on the Blackfin architecture led to the discovery of an
incomplete definition of the RO_DATA macro used in this series.  The fixes
are in linux-next:

	commit 906f2a51c941 ("mm: fix section name for .data..ro_after_init")

	commit 939897e2d736 ("vmlinux.lds: add missing VMLINUX_SYMBOL macros")

This patch (of 2):

Implement a mechanism to check if a module's address is in the rodata or
ro_after_init sections.  It mimics the existing functions that test if an
address is inside a module's text section.

Functions that take a module as an argument will be able to verify that
the module address is in a read-only section.  The idea is to prevent
structures (including modules) that are not read-only from being passed to
functions.

This implements the first half of a suggestion made by Kees Cook for the
Kernel Self Protection Project:

    - provide mechanism to check for ro_after_init memory areas, and
      reject structures not marked ro_after_init in vmbus_register()

Link: http://lkml.kernel.org/r/20170406033550.32525-2-ewk@edkovsky.org
Signed-off-by: Eddie Kovsky <ewk@edkovsky.org>
Suggested-by: Kees Cook <keescook@chromium.org>
Acked-by: Jessica Yu <jeyu@redhat.com>
Acked-by: Kees Cook <keescook@chromium.org>
Cc: Rusty Russell <rusty@rustcorp.com.au>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>