extable: verify address is read-only
Provide a mechanism to check if the address of a variable is const or
ro_after_init.  It mimics the existing functions that test if an address
is inside the kernel's text section.

The idea is to prevent structures that are not read-only from being passed
to functions.  Other functions inside the kernel could then use this
capability to verify that their arguments are read-only.

This implements the first half of a suggestion made by Kees Cook for the
Kernel Self Protection Project:

	- provide mechanism to check for ro_after_init memory areas, and
	reject structures not marked ro_after_init in vmbus_register()

Link: http://lkml.kernel.org/r/20170406033550.32525-3-ewk@edkovsky.org
Signed-off-by: Eddie Kovsky <ewk@edkovsky.org>
Suggested-by: Kees Cook <keescook@chromium.org>
Acked-by: Kees Cook <keescook@chromium.org>
Cc: Jessica Yu <jeyu@redhat.com>
Cc: Rusty Russell <rusty@rustcorp.com.au>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>