sched: remove sched_find_first_bit()
sched_find_first_bit() is in fact the unrolled version of
find_first_bit(), which is theoretically faster in some cases.
But in the kernel it is called only in couple places in
kernel/sched/rt.c, and both of them are not looking like hot
paths that will achieve any measurable benefit from using unrolled
version of find_first_bit() - there's no hard loops, and the
execution path is not really short.

This patch removes sched_find_first_bit() and deletes
include/asm-generic/bitops/sched.h as it declares only this function.
Alpha has it's own implementation, so it is also removed.

Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>