arch: remove GENERIC_FIND_FIRST_BIT
This option, if disabled, is used to alias find_first_bit() to
find_next_bit() with the trivial macro:

And the same for find_first_zero_bit().

The problem here is that the implementation of find_next_bit() is more
complex, and that extra complexity is not really needed if the offset
that passed to find_next_bit() is known to be 0.

This patch removes GENERIC_FIND_FIRST_BIT and drops the alias to find_next_bit().
Architectures that enable GENERIC_FIND_FIRST_BIT will be obviously not affected
with this change. Namely, arc, s390, tile, unicore32 and x86. Some architectures
implement their own implementations, so they are not affected too. They are: arm,
m68k, unicore32. (Unicore32 both enables the CONFIG_GENERIC_FIND_FIRST_BIT and has
custom implementation for it, so it fixed too.) Others will switch to separate
implementation.

Tested on arm64.

There is a couple of patches in the kernel that remove similar config options:
63e424c84429 ("arch: remove CONFIG_GENERIC_FIND_{NEXT_BIT,BIT_LE,LAST_BIT}") and
65af3a3f89d7 ("arch: remove CONFIG_GENERIC_FIND_NEXT_BIT again"), but the
GENERIC_FIND_FIRST_BIT is the last one that still there. So this path finishes
the work.

Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>