bitmap: new bitmap_copy_safe and bitmap_{from,to}_arr32
'safe' in bitmap_copy_safe stands for clearing tail bits in bitmap
beyond last bit till the end of last word. It is useful for hardening
API when bitmap is assumed to be provided to userspace.

bitmap_{from,to}_arr32 functions are replacements for bitmap_{from,to}_u32array.
They don't take unneeded nwords argument, and so simpler in implementation
and understanding.

This patch suggests optimization for 32-bit systems - aliasing
bitmap_{from,to}_arr32 to bitmap_copy_safe. Other possible optimization
is to map 64-bit LE bitmap_{from,to}_arr32 to more generic function(s).

But I didn't end up with the function that would be helpful by itself,
and allow alias 64-bit LE bitmap_{from,to}_arr32, like bitmap_copy_safe
does. So I did prefer to leave things as is.

Since this is RFC, I would like to ask people here for ideas. :)